V_1.0.6

+++ src/main/resources/db/migration/V01__criar_tabela_paciente.sql
@@ -1,6 +1,6 @@
 CREATE TABLE paciente(
-	cpf INTEGER PRIMARY KEY,
-	rg INTEGER,
+	cpf BIGINT PRIMARY KEY,
+	rg BIGINT,
 	nome VARCHAR(255),
 	endereco VARCHAR(255),
 	numero_casa INTEGER,
